Enhance Maker Fidelity UX
This includes:
 - use better delay logic while waiting for any funding tx to be in mempool and confirming the fidelity tx.
 - use `get_next_external_address` for recieving any funds instead of `fidelity` address.
 - sync the wallet inside the `create_fidelity` api and save in-memory data while shutting down the maker server.
 - show balance in `BTC` as  bitcoin-cli wallet defaults to `btc` amounts.

+++ src/wallet/error.rs
@@ -17,7 +17,7 @@ pub enum WalletError {
     Locktime(bitcoin::blockdata::locktime::absolute::ConversionError),
     Secp(bitcoin::secp256k1::Error),
     Consensus(String),
-    InsufficientFund { available: u64, required: u64 },
+    InsufficientFund { available: f64, required: f64 },
 }
 
 impl From<std::io::Error> for WalletError {
